Log Java ZXing library version if decoding fails
The Java ZXing library contains a fair number of version-specific gaps and
quirks.

For example, prior to the patch I contributed in
https://github.com/zxing/zxing/commit/20405981dcc2f44e80bf90c6b25e8aaf6066e93f
(which is not yet in a release), ZXing would fail to decode PDF417 barcodes
in vertical orientations, which are very common in airline boarding passes.

So it's useful to log the ZXing library version in order to address users'
questions about differences between different ZXing decoder interfaces.