Handle raw_bits==None correctly in repr()
This should fix #36.

Also adds tests to verify that `BarCode.parse()` results in a working
`__repr__`, whether or not the "Raw bits:" section appears in the ZXing CLR
output.

However, I can't actually reproduce this condition using the "normal" entry
point of `BarCodeReader.decode()` → `BarCode.parse()` on the latest version.